Fix TrackedReferences naming
