+++ src/Umbraco.Cms.ManagementApi/Controllers/TrackedReference/ItemsTrackedReferenceController.cs
@@ -7,33 +7,32 @@ using Umbraco.Cms.ManagementApi.ViewModels.Pagination;
 using Umbraco.Cms.ManagementApi.ViewModels.TrackedReferences;
 using Umbraco.New.Cms.Core.Models;
 
-namespace Umbraco.Cms.ManagementApi.Controllers.TrackedReferences;
+namespace Umbraco.Cms.ManagementApi.Controllers.TrackedReference;
 
-public class DescendantsTrackedReferencesController : TrackedReferencesControllerBase
+public class ItemsTrackedReferenceController : TrackedReferenceControllerBase
 {
     private readonly ITrackedReferencesService _trackedReferencesSkipTakeService;
     private readonly IUmbracoMapper _umbracoMapper;
 
-    public DescendantsTrackedReferencesController(ITrackedReferencesService trackedReferencesSkipTakeService, IUmbracoMapper umbracoMapper)
+    public ItemsTrackedReferenceController(ITrackedReferencesService trackedReferencesSkipTakeService, IUmbracoMapper umbracoMapper)
     {
         _trackedReferencesSkipTakeService = trackedReferencesSkipTakeService;
         _umbracoMapper = umbracoMapper;
     }
 
     /// <summary>
-    ///     Gets a page list of the child nodes of the current item used in any kind of relation.
+    ///     Gets a page list of the items used in any kind of relation from selected integer ids.
     /// </summary>
     /// <remarks>
-    ///     Used when deleting and unpublishing a single item to check if this item has any descending items that are in any
-    ///     kind of relation.
-    ///     This is basically finding the descending items which are children in relations.
+    ///     Used when bulk deleting content/media and bulk unpublishing content (delete and unpublish on List view).
+    ///     This is basically finding children of relations.
     /// </remarks>
-    [HttpGet("descendants/{parentId:int}")]
+    [HttpGet("item")]
     [MapToApiVersion("1.0")]
     [ProducesResponseType(typeof(PagedViewModel<RelationItemViewModel>), StatusCodes.Status200OK)]
-    public async Task<ActionResult<PagedViewModel<RelationItemViewModel>>> Descendants(int parentId, long skip, long take, bool? filterMustBeIsDependency)
+    public async Task<ActionResult<PagedViewModel<RelationItemViewModel>>> GetPagedReferencedItems([FromQuery]int[] ids, long skip, long take, bool? filterMustBeIsDependency)
     {
-        PagedModel<RelationItemModel> relationItems = _trackedReferencesSkipTakeService.GetPagedDescendantsInReferences(parentId, skip, take, filterMustBeIsDependency ?? true);
+        PagedModel<RelationItemModel> relationItems = _trackedReferencesSkipTakeService.GetPagedItemsWithRelations(ids, skip, take, filterMustBeIsDependency ?? true);
         var pagedViewModel = new PagedViewModel<RelationItemViewModel>
         {
             Total = relationItems.Total,
