+++ src/Umbraco.Cms.ManagementApi/Controllers/TrackedReference/TrackedReferencesControllerBase.cs
@@ -2,12 +2,12 @@
 using NSwag.Annotations;
 using Umbraco.New.Cms.Web.Common.Routing;
 
-namespace Umbraco.Cms.ManagementApi.Controllers.TrackedReferences;
+namespace Umbraco.Cms.ManagementApi.Controllers.TrackedReference;
 
 [ApiController]
 [VersionedApiBackOfficeRoute("tracked-reference")]
 [OpenApiTag("Tracked Reference")]
 [ApiVersion("1.0")]
-public abstract class TrackedReferencesControllerBase : ManagementApiControllerBase
+public abstract class TrackedReferenceControllerBase : ManagementApiControllerBase
 {
 }
